# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "The books managing API"
  },
  "tags": [
    {
      "name": "Books",
      "description": "The books managing API"
    }
  ],
  "paths": {
    "api/books": {
      "get": {
        "summary": "Lists all the books",
        "tags": ["Books"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "Search keyword to filter books by title or author"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["createdAt","title","finished"]
            },
            "required": false,
            "description": "Field to sort by (title or createdAt)"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"]
            },
            "required": false,
            "description": "Sort order direction"
          }
        ],
        "responses": {
          "200": {
            "description": "The list of the books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "400":{
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new book",
        "tags": ["Books"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              },
            "example": {
                  "title": "Clean Code",
                  "author": "Robert C. Martin",
                  "finished": false
            }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created book.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "409": {
            "description": "Book already exists"
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "api/books/{id}": {
      "get": {
        "summary": "Get the book by id",
        "tags": ["Books"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The book id"
          }
        ],
        "responses": {
          "200": {
            "description": "The book response by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "The book was not found"
          }
        }
      },
      "put": {
        "summary": "Update the book by the id",
        "tags": ["Books"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The book id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The book was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "404": {
            "description": "The book was not found"
          },
          "500": {
            "description": "Some error happened"
          }
        }
      },
      "delete": {
        "summary": "Remove the book by id",
        "tags": ["Books"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The book id"
          }
        ],
        "responses": {
          "200": {
            "description": "The book was deleted"
          },
          "404": {
            "description": "The book was not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "required": ["title", "author", "finished"],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the book"
          },
          "title": {
            "type": "string",
            "description": "The title of your book"
          },
          "author": {
            "type": "string",
            "description": "The book author"
          },
          "finished": {
            "type": "boolean",
            "description": "Whether you have finished reading the book"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "description": "The date the book was added"
          }
        },
        "example": {
          "title": "The New Turing Omnibus",
          "author": "Alexander K. Dewdney",
          "finished": false,
          "createdAt": "2025-04-10T04:05:06.157Z"
        }
      }
    }
  }
}